New updearter
This is the latest update file for the Japanese version of Pro bot. Install this on your bot via your computer. The installation procedure is to prepare your own token on the server, run /updeaterregist and upload your files there. You will then receive a direct message. Authenticate it and it will be installed.